+++ src/main/java/pl/plajer/villagedefense3/events/JoinEvent.java
@@ -85,11 +85,6 @@ public class JoinEvent implements Listener {
     try {
       //we want to be the first :)
       Bukkit.getScheduler().runTaskLater(plugin, () -> {
-        if (event.getPlayer().isOp() && !plugin.isDataEnabled()) {
-          event.getPlayer().sendMessage(ChatColor.RED + "[VillageDefense] It seems that you've disabled bStats statistics.");
-          event.getPlayer().sendMessage(ChatColor.RED + "Please consider enabling it to help us develop our plugins better!");
-          event.getPlayer().sendMessage(ChatColor.RED + "Enable it in plugins/bStats/config.yml file");
-        }
         if (event.getPlayer().hasPermission("villagedefense.updatenotify")) {
           if (plugin.getConfig().getBoolean("Update-Notifier.Enabled", true)) {
             String currentVersion = "v" + Bukkit.getPluginManager().getPlugin("VillageDefense").getDescription().getVersion();